Created MORE_OFF promotion. Changed freeItem field from Integer to Double. Changed sql script according to PromotionConfig class. Optimized PriceHandler class to work with Double variable and added Double values to freeItems.

+++ src/main/java/dev/aziz/grocerystore/entities/PromotionConfig.java
@@ -33,7 +33,7 @@ public class PromotionConfig {
     private Integer minimumAmount;
 
     @NotNull(message = "Free amount should not be empty")
-    private Integer freeAmount;
+    private Double freeAmount;
 
     @OneToOne
     private Item item;
